chat_service/requirements.txt
fastapi==0.104.1
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.2.2
pymongo==4.5.0
numpy==1.26.0
pandas==2.1.1
requests==2.31.0
python-dotenv==1.0.0
